// frontend/src/api/model/startRegistrationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Project HOME
 * 백엔드 멀티 모듈 API 명세서
 * OpenAPI spec version: 1.0
 */

export interface StartRegistrationResponse {
  username?: string;
  rpId?: string;
  challenge?: string;
  userId?: string;
  displayName?: string;
  pubKeyCredParams?: string;
}